requirements.txt
numpy
pygame
pyglet
PyWavefront
PyGLM
./PyOpenGL-3.1.5-cp39-cp39-win_amd64.whl
./PyOpenGL_accelerate-3.1.5-cp39-cp39-win_amd64.whl